edit folder structure
