+++ src/components/Dashboard/Dashboard.js
@@ -1,9 +1,9 @@
-import { Box, Divider, Flex, Heading } from '@chakra-ui/react';
-import React from 'react';
+import { Divider, Flex } from "@chakra-ui/react";
+import React from "react";
 import { useNavigate } from "react-router-dom";
-import { CategoryTable } from '../../elements/menuElements';
-import Header from '../Header/Header';
-import menu from "./Dashboard.constants";
+import Header from "../Header/Header";
+import Menu from "../Menu/Menu";
+
 const Dashboard = () => {
   const navigate = useNavigate();
   const user = JSON.parse(localStorage.getItem("loggedInUser"));
@@ -16,16 +16,7 @@ const Dashboard = () => {
     <Flex direction="column" gap="12px" paddingBottom="20px">
       <Header />
       <Divider />
-      <Box padding="20px">
-        <Heading fontFamily="'Helvetica', sans-serif" color="#5F7161">
-          Menu
-        </Heading>
-        <Flex direction="column" flexWrap="wrap" gap="16px" alignItems="center">
-          {menu.categories.map((category, index) => (
-            <CategoryTable key={index} category={category} />
-          ))}
-        </Flex>
-      </Box>
+      <Menu />
     </Flex>
   );
 };
