+++ src/components/LogInForm/LogInForm.js
@@ -2,7 +2,9 @@ import React from 'react';
 import { useDispatch } from "react-redux";
 import { useNavigate } from 'react-router-dom';
 import SubmitButton from "../../common/buttons/SubmitButton.js";
-import { FormContainer, PasswordInput, TextInput } from "../../elements/formElements";
+import FormContainer from "../../common/forms/FormContainer";
+import PasswordInput from "../../common/forms/PasswordInput";
+import TextInput from "../../common/forms/TextInput";
 import useField from '../../hooks/useField';
 import { logUserIn } from '../../reducers/loggedInUserReducer';
 import loginService from '../../services/login';
