+++ src/components/Menu/index.js
@@ -1,0 +1,2 @@
+export * from "./Menu";
+export { default } from "./Menu";
